+++ projects/client/src/views/OrdersList/components/ordersList.jsx
@@ -26,7 +26,7 @@ export const OrdersList = () => {
 	};
 	const ordersList = async () => {
 		try {
-			const response = await Axios.get(`${process.env.REACT_APP_API_BASE_URL}/order`, { headers });
+			const response = await Axios.get(`${process.env.REACT_APP_API_BASE_URL}/user/orders`, { headers });
 			setList(response.data.result);
 			console.log(response.data);
 		} catch (error) {
