+++ projects/server/src/controllers/index.js
@@ -2,12 +2,10 @@ const userControllers = require('./userControllers');
 const adminControllers = require('./adminControllers');
 const productControllers = require('./productControllers');
 const categoryControllers = require('./categoryControllers');
-const orderControllers = require('./orderControllers');
 
 module.exports = {
     userControllers,
     adminControllers,
     productControllers,
     categoryControllers,
-    orderControllers
 };
