--- projects/server/src/controllers/orderControllers.js
@@ -1,41 +1,0 @@
-const db = require("../models");
-const orders = db.Orders;
-const order_details = db.Order_details;
-const products = db.Products;
-const { Op } = require("sequelize");
-
-module.exports = {
-	ordersList: async (req, res) => {
-		try {
-			const page = +req.query.page || 1;
-			const limit = +req.query.limit || 4;
-			const search = req.query.search;
-			const offset = (page - 1) * limit;
-			const condition = {};
-			if (search) {
-				condition[Op.or] = [
-					{
-						productName: {
-							[Op.like]: `%${search}%`,
-						},
-					},
-				];
-			}
-			const result = await order_details.findAll({
-				include: [{ model: orders }, { model: products, where: condition }],
-				limit,
-				offset,
-			});
-			res.status(200).send({
-				result,
-				currentPage: page,
-			});
-		} catch (error) {
-			return res.status(500).send({
-				error,
-				status: 500,
-				message: "Internal server error.",
-			});
-		}
-	},
-};
