+++ projects/server/src/routers/index.js
@@ -2,12 +2,10 @@ const userRouters = require('./userRouters');
 const adminRouters = require('./adminRouters');
 const productRouters = require('./productRouters');
 const categoryRouters = require('./categoryRouters');
-const orderRouters = require('./orderRouters');
 
 module.exports = {
     userRouters,
     adminRouters,
     productRouters,
     categoryRouters,
-    orderRouters
 };
